+++ app/views/press_catalog/_thumbnail_monograph.html.erb
@@ -1,5 +1,5 @@
 <% if has_thumbnail?(document) && tn = render_thumbnail_tag(document, {}, counter: document_counter_with_offset(document_counter)) %>
   <div class="col-sm-3">
-    <a href="<%= polymorphic_path(url_for_document(document)) %>" ><%= raw presenter.monograph_thumbnail(225) %></a>
+    <a href="<%= polymorphic_path(url_for_document(document)) %>" ><%= raw presenter.work_thumbnail(225) %></a>
   </div>
 <% end %>
